+++ plan.md
@@ -1,3 +1,4 @@
 Feature-Sliced Design に従ってディレクトリ構成を変更する。
 
 - [x] src/shared/api/useGetPostsForUser.tsx を src/entities/post/api へ移動する
+- [x] src/shared/api/useGetTimelinePosts.tsx を src/entities/post/api へ移動する
