+++ src/entities/post/index.ts
@@ -1,1 +1,2 @@
 export { useGetPostsForUser } from "./api/useGetPostsForUser";
+export { useGetTimelinePosts } from "./api/useGetTimelinePosts";
