+++ src/shared/api/index.ts
@@ -1,2 +1,1 @@
-export { useGetTimelinePosts } from "./useGetTimelinePosts";
 export { useSearchPosts } from "./useSearchPosts";
